# WC.Admin/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WC.API",
    "version": "1.0"
  },
  "paths": {
    "/Country": {
      "get": {
        "tags": [
          "Country"
        ],
        "operationId": "GetCountries",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryViewModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryViewModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryViewModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Country/{search}": {
      "get": {
        "tags": [
          "Country"
        ],
        "operationId": "GetCountriesFiltered",
        "parameters": [
          {
            "name": "search",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryViewModel"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryViewModel"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CountryViewModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Dashboard": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "operationId": "GetDashboardSummary",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/Imports": {
      "post": {
        "tags": [
          "Imports"
        ],
        "operationId": "ImportIpRanges",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportViewModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ImportViewModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ImportViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportViewModel"
                }
              }
            }
          }
        }
      }
    },
    "/IpRange": {
      "get": {
        "tags": [
          "IpRange"
        ],
        "operationId": "GetIpRanges",
        "parameters": [
          {
            "name": "CountryId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "IpVersion",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "ActiveOnly",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IpRangePagedResultModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IpRangePagedResultModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IpRangePagedResultModel"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "IpRange"
        ],
        "operationId": "CreateIpRange",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IpRangeViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IpRangeViewModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/IpRangeViewModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/IpRange/{id}": {
      "get": {
        "tags": [
          "IpRange"
        ],
        "operationId": "GetIpRangeById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IpRangeViewModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IpRangeViewModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IpRangeViewModel"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "IpRange"
        ],
        "operationId": "UpdateIpRange",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IpRangeViewModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IpRangeViewModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/IpRangeViewModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "IpRange"
        ],
        "operationId": "DeleteIpRange",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/IpRange/GetCountryFromIpAdress": {
      "get": {
        "tags": [
          "IpRange"
        ],
        "operationId": "GetCountryFromIpAdress",
        "parameters": [
          {
            "name": "ipAddress",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CountryResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountryResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CountryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/IpRange/RefreshIpIntegersBinariesAndVersions": {
      "get": {
        "tags": [
          "IpRange"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/IpRange/RefreshIPv6HighsAndLows": {
      "get": {
        "tags": [
          "IpRange"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CountryResponse": {
        "type": "object",
        "properties": {
          "countryName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CountryViewModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "iso2Code": {
            "type": "string",
            "nullable": true
          },
          "iso3Code": {
            "type": "string",
            "nullable": true
          },
          "ipRangeCount": {
            "type": "integer",
            "format": "int32"
          },
          "activeIpRangeCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "DashboardViewModel": {
        "type": "object",
        "properties": {
          "totalCountries": {
            "type": "integer",
            "format": "int32"
          },
          "totalIpRanges": {
            "type": "integer",
            "format": "int32"
          },
          "activeIpRanges": {
            "type": "integer",
            "format": "int32"
          },
          "iPv4Ranges": {
            "type": "integer",
            "format": "int32"
          },
          "iPv6Ranges": {
            "type": "integer",
            "format": "int32"
          },
          "lastImportDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastImportStatus": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ImportViewModel": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "nullable": true
          },
          "resultMessage": {
            "type": "string",
            "nullable": true
          },
          "success": {
            "type": "boolean",
            "nullable": true
          },
          "processedCount": {
            "type": "integer",
            "format": "int32"
          },
          "insertedCount": {
            "type": "integer",
            "format": "int32"
          },
          "updatedCount": {
            "type": "integer",
            "format": "int32"
          },
          "skippedCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "IpRangePagedResultModel": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IpRangeViewModel"
            },
            "nullable": true
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "countryId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ipVersion": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "activeOnly": {
            "type": "boolean"
          },
          "search": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "IpRangeViewModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "countryId": {
            "type": "integer",
            "format": "int32"
          },
          "countryName": {
            "type": "string",
            "nullable": true
          },
          "countryCodeIso2": {
            "type": "string",
            "nullable": true
          },
          "ipVersion": {
            "type": "integer",
            "format": "int32"
          },
          "startIp": {
            "type": "string",
            "nullable": true
          },
          "endIp": {
            "type": "string",
            "nullable": true
          },
          "active": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    }
  }
}